+++ tests/run-nunit-tests.sh
@@ -52,7 +52,7 @@ PROJECTS=(
 declare -A EXPECTED_DISCOVERED=(
 	["tests/basic"]=590
 	["tests/coroutines"]=221
-	["tests/roundtrip/consumer"]=204
+	["tests/roundtrip/consumer"]=212
 	["tests/roundtrip/bidirectional/consumer"]=37
 	["tests/interop/consumer"]=179
 )